removing old example files
