+++ packages/example/package.json
@@ -15,7 +15,6 @@
     "@colyseus/uwebsockets-transport": "^0.15.0",
     "@colyseus/ws-transport": "^0.15.0",
     "@colyseus/redis-driver": "^0.15.0",
-    "@colyseus/redis-presence": "^0.15.0",
-    "@colyseus/mongoose-driver": "^0.15.0"
+    "@colyseus/redis-presence": "^0.15.0"
   }
 }
