+++ packages/example/src/MyRoom.ts
@@ -1,76 +1,32 @@
-import { Room, Client } from "colyseus";
-import { InputData, MyRoomState, Player } from "./schema/TestState";
+import { Room, Client, ClientArray } from "colyseus";
+import { Schema, type, MapSchema } from "@colyseus/schema";
+import { IncomingMessage } from "http";
+
+export class Player extends Schema {
+  @type("number") x: number;
+  @type("number") y: number;
+}
+
+export class MyRoomState extends Schema {
+  @type("number") mapWidth: number;
+  @type("number") mapHeight: number;
+  @type({ map: Player }) players = new MapSchema<Player>();
+}
 
 export class MyRoom extends Room<MyRoomState> {
-  fixedTimeStep = 1000 / 60;
-
   players: { [sessionId: string]: any } = {};
 
   onCreate(options: any) {
     this.setState(new MyRoomState());
 
-    // set map dimensions
+    // map dimensions
     this.state.mapWidth = 800;
     this.state.mapHeight = 600;
 
-    this.onMessage(0, (client, input) => {
+    this.onMessage('input', (client, input) => {
       // handle player input
       const player = this.state.players.get(client.sessionId);
-
-      // enqueue input to user input buffer.
-      player.inputQueue.push(input);
-    });
-
-    this.onMessage("hello", (client, input) => {
-      // handle player input
-      const player = this.state.players.get(client.sessionId);
-      player.x = Number(input);
-
-      // enqueue input to user input buffer.
-      player.inputQueue.push(input);
-    });
-
-    this.onMessage("new", (client, input) => {
-      // handle player input
-      const player = this.state.players.get(client.sessionId);
-      player.something = Number(input);
-    });
-
-    let elapsedTime = 0;
-    this.setSimulationInterval((deltaTime) => {
-      elapsedTime += deltaTime;
-
-      while (elapsedTime >= this.fixedTimeStep) {
-        elapsedTime -= this.fixedTimeStep;
-        this.fixedTick(this.fixedTimeStep);
-      }
-    });
-  }
-
-  fixedTick(timeStep: number) {
-    const velocity = 2;
-
-    this.state.players.forEach(player => {
-      let input: InputData;
-
-      // dequeue player inputs
-      while (input = player.inputQueue.shift()) {
-        if (input.left) {
-          player.x -= velocity;
-
-        } else if (input.right) {
-          player.x += velocity;
-        }
-
-        if (input.up) {
-          player.y -= velocity;
-
-        } else if (input.down) {
-          player.y += velocity;
-        }
-
-        player.tick = input.tick;
-      }
+      player.x += 1; //dummy mutation
     });
   }
 
@@ -95,10 +51,6 @@ export class MyRoom extends Room<MyRoomState> {
 
   onRestoreRoom(cached: any): void {
     console.log("ROOM HAS BEEN RESTORED!", cached);
-
-    this.state.players.forEach(player => {
-      player.method();
-    });
   }
 
   onDispose() {
