+++ packages/example/src/ServerTCP.ts
@@ -5,10 +5,12 @@
 import * as net from "net";
 
 import { Server } from "@colyseus/core";
-import { DummyRoom } from "./DummyRoom";
+import { MyRoom } from "./MyRoom";
 
 const port = Number(process.env.PORT || 8181);
 const endpoint = "localhost";
+
+/*
 
 const engine = net.Server;
 
@@ -16,9 +18,11 @@ const engine = net.Server;
 const gameServer = new Server({ engine });
 
 // Register DummyRoom as "chat"
-gameServer.define("chat", DummyRoom);
+gameServer.define("my_room", MyRoom);
 
 process.on('unhandledRejection', r => console.log(r));
 gameServer.listen(port);
 
 console.log(`Listening on tcp://${ endpoint }:${ port }`)
+
+*/
